Commit pending changes before trying another approach
The main issue with the current one is alpha blending. It's not possible to
render the whole hierarchy into a single texture because semi-transparent
children override alpha written by their fully-opaque parents. It probably
would be fixed by separate alpha blending, but it requires changing of the
shader which makes this approach unappealing.